# Conflicts:
#	BACKEND/Data/Migrations/RecruitmentWebContextModelSnapshot.cs
#	BackEnd/Api/appsettings.json
#	BackEnd/Service/InterviewService.cs
#	FRONTEND/src/app/module/cv/cv-manage/upload-cv/upload-cv.component.ts
#	FrontEnd/package-lock.json
#	FrontEnd/src/app/app.routes.ts
#	FrontEnd/src/app/module/position/position.component.html
#	FrontEnd/src/app/module/position/position.component.ts

+++ BackEnd/Service/EventService.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Data.CustomModel.Event;
 using Data.Entities;
 using Data.Interfaces;
 using Microsoft.IdentityModel.Tokens;
@@ -27,27 +28,58 @@ public class EventService : IEventService
 
     public async Task<bool> DeleteEvent(Guid eventModelId)
     {
+        var foundEvent = await this.GetEventById(eventModelId, true);
+        if (foundEvent == null)
+        {
+            return await Task.FromResult(false);
+        }
+        if (foundEvent.StartDateTime <= DateTime.Now && DateTime.Now <= foundEvent.EndDateTime)
+        {
+            return await Task.FromResult(false);
+        }
         return await _EventRepository.DeleteEvent(eventModelId);
     }
 
-    public async Task<IEnumerable<EventModel>> GetAllEvent()
+    public async Task<IEnumerable<EventModel>> GetAllEvent(bool isAdmin)
     {
         var data = await _EventRepository.GetAllEvent();
         if (!data.IsNullOrEmpty())
         {
             List<EventModel> result = _mapper.Map<List<EventModel>>(data);
-            return result;
+             
+            return !isAdmin ? result.Where(e => !e.IsDeleted) : result;
+        }
+        return null!;
+    }
+
+    public async Task<IEnumerable<EventModel>> GetAllEventByRecruiterId(Guid recruiterId, EventFilter eventFilter, string sortString, bool isAdmin)
+    {
+        var data = await _EventRepository.GetAllEventByRecruiterId(recruiterId, eventFilter, sortString);
+       
+        if (!data.IsNullOrEmpty())
+        {
+            List<EventModel> result = _mapper.Map<List<EventModel>>(data);
+            return isAdmin ? result : result.Where(o => !o.IsDeleted).ToList();
         }
         return null!;
     }
 
     public async Task<bool> UpdateEvent(EventModel eventModel, Guid eventModelId)
     {
+        var foundEvent = await this.GetEventById(eventModelId, true);
+        if (foundEvent == null)
+        {
+            return await Task.FromResult(false);
+        }
+        if (foundEvent.StartDateTime <= DateTime.Now && DateTime.Now <= foundEvent.EndDateTime)
+        {
+            return await Task.FromResult(false);
+        }
         var data = _mapper.Map<Event>(eventModel);
         return await _EventRepository.UpdateEvent(data, eventModelId);
     }
 
-    public async Task<EventModel> GetEventById(Guid id)
+    public async Task<EventModel> GetEventById(Guid id, bool isAdmin)
     {
         var data = await _EventRepository.GetEventById(id);
         if (data == null)
@@ -55,6 +87,10 @@ public class EventService : IEventService
             return null!;
         }
         var respone = _mapper.Map<EventModel>(data);
+        if (!isAdmin)
+        {
+            return respone.IsDeleted ? null : respone;
+        }
         return respone;
     }
 }
